+++ src/main/kotlin/in/kkkev/jjidea/util/Tasks.kt
@@ -2,7 +2,17 @@ package `in`.kkkev.jjidea.util
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ModalityState
+import com.intellij.openapi.application.TransactionGuard
+import com.intellij.openapi.fileEditor.FileDocumentManager
 import java.awt.Component
+
+fun saveAllDocuments() {
+    if (TransactionGuard.getInstance().isWriteSafeModality(ModalityState.current())) {
+        ApplicationManager.getApplication().runWriteIntentReadAction<Unit, Nothing> {
+            FileDocumentManager.getInstance().saveAllDocuments()
+        }
+    }
+}
 
 private val capturedModality = ThreadLocal<ModalityState>()